+++ src/widgets/list-with-input-panel/containers/list-with-input-panel-container.jsx
@@ -34,8 +34,6 @@ const propTypes = {
   validateForm: PropTypes.func.isRequired,
 
   resetObject: PropTypes.object.isRequired,
-
-  isQuestionnaireModified: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -54,7 +52,6 @@ const mapStateToProps = (state, { formName, selectorPath }) => {
     componentId: state.appState.editingComponentId,
     formValues: getFormValues(formName)(state),
     currentValues: selector(state, selectorPath),
-    isQuestionnaireModified: state.appState.isQuestionnaireModified,
   };
 };